Move generate options out of dialog
